Moved configuration to Spring
Application configuration is now a Spring Bean.
